Default to a signal value instead of None
Example:

```
experiment_id = 1
conn.experiments(experiment_id).fetch()
```

This will call `/v1/experiments/1`

If however, `experiment_id` is accidentally `None`, then the call
will be

```
/v1/experiments
```

which has a completely different format which will cause subtle
errors. With this change, the call will be

```
/v1/experiments/None
```

which is not a real endpoint, but the error is more apparent.

I think this is the right behaviour since it's unlikely that someone
will actually want to switch programmatically between
`conn.experiments(1)` and `conn.experiments()`

@alexandraj777