Fix mutability of ApiObjects
If you had an Experiment object `e = Experiment({'name': 'abc'})`, and did

```
print e.name
e.name = 'def'
print e.name
```

It would print "abc" and then "def". However, this would be overwriting
the `name = Field(str)` with just `name = 'def'`, which is not what
we want. Furthermore, `e.to_json() == {'name': 'abc'}`. This behaviour
is unexpected and so this resolves this behaviour.

If instead we wanted the ApiObjects to be immutable, we could do that
as well, though that would be backwards-incompatible, and I think that
there are reasonable uses for mutating objects. If users want to treat
them as immutable that is their decision, and not one I think we need
to enforce.